+++ finch/processes/base.py
@@ -1,3 +1,5 @@
+from dask.diagnostics import ProgressBar
+from dask.diagnostics.progress import format_time
 from pywps import Process
 from sentry_sdk import configure_scope
 import xarray as xr
@@ -78,3 +80,21 @@ def chunk_dataset(ds, max_size=1000000):
         chunks[dim] = max(chunks[dim] // 2, 1)
 
     return chunks
+
+
+class FinchProgress(ProgressBar):
+    def __init__(self, logging_function, start_percentage, *args, **kwargs):
+        super(FinchProgress, self).__init__(*args, **kwargs)
+        self._logging_function = logging_function
+        self._start_percentage = start_percentage
+
+    def _draw_bar(self, frac, elapsed):
+        start = self._start_percentage / 100
+
+        frac += start - frac * start
+        bar = '#' * int(self._width * frac)
+        percent = int(100 * frac)
+        elapsed = format_time(elapsed)
+        msg = '[{0:<{1}}] | {2}% Done | {3}'.format(bar, self._width, percent, elapsed)
+
+        self._logging_function(msg, percent)
